LPS-71620 Move
